Fix setup of repository storages for the test environment
On the backup_rake_spec we require a second repository storage, which we
used to accomplish through a combination of rspec-mocks and special
configurations in the Gitaly installation. Since Gitaly now refuses to
start with invalid storages, we modified the setup code in
https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/22530, but in
between tests the storage could be removed from the `tmp/` folder.

These changes simplify the approach by making the secondary storage a
_real_ storage and having `TestEnv` manage it the same way it does the
default storage.