Added load_content option to avoid loading all wiki pages content
Inside a wiki, when we show the sidebar or browse to the `pages`,
all page contents are retrieved from Gitaly and that is a waste
of resources, since no content from that pages are going to be
showed.

This MR introduces the param `load_content` in `ProjectWiki#pages`,
to enable/disabled retrieving that content.

Also in the `WikisController` we disable retrieving the content
from pages in the sidebar and also the ones showed in the `pages`
page.